order_service/requirements.txt
djangorestframework
django-environ
gunicorn
psycopg2-binary
djangorestframework-simplejwt
python-dotenv
requests
pika
kafka-python
line_profiler
httpx